+++ R/generate_matrices.R
@@ -273,6 +273,15 @@ generate_matrix_SBS <- function(query, ref_genome, genome_build = "hg19", add_tr
       APOBEC_scores = sub.tbl
     )
   }
+
+  # Reorder mutation types
+  res$all_matrices = lapply(res$all_matrices, function(x) {
+    y = x[, sort(colnames(x))]
+    y
+  })
+  res$nmf_matrix = res$nmf_matrix[, sort(colnames(res$nmf_matrix))]
+
+  # Return
   res
 }
 
@@ -404,6 +413,14 @@ generate_matrix_DBS <- function(query, ref_genome, genome_build = "hg19", add_tr
       )
     )
   }
+
+  # Reorder mutation types
+  res$all_matrices = lapply(res$all_matrices, function(x) {
+    y = x[, sort(colnames(x))]
+    y
+  })
+  res$nmf_matrix = res$nmf_matrix[, sort(colnames(res$nmf_matrix))]
+
   res
 }
 
@@ -657,6 +674,13 @@ generate_matrix_INDEL <- function(query, ref_genome, genome_build = "hg19", add_
       )
     )
   }
+  # Reorder mutation types
+  res$all_matrices = lapply(res$all_matrices, function(x) {
+    y = x[, sort(colnames(x))]
+    y
+  })
+  res$nmf_matrix = res$nmf_matrix[, sort(colnames(res$nmf_matrix))]
+
   res
 }